+++ docs/README.md
@@ -101,9 +101,19 @@ To help you convert HTML snippets to Elm code and help newcomers learn the synta
 
 __Setting:__ `elmLand.feature.goToSymbol`
 
-You can navigate symbols inside a file. This is helpful for quickly navigating among functions, values and types in a file. You can also navigate to any top-level declaration in any file, which is a quick way of getting to the right file.
+You can navigate symbols inside a file. This is helpful for quickly navigating among functions, values and types in a file. The Outline panel below the file tree in the sidebar also displays all functions, values and types in the file.
 
 ![Go to symbol](./go-to-symbol.gif)
+
+---
+
+### __Open symbol by name__
+
+__Setting:__ `elmLand.feature.openSymbolByName`
+
+You can navigate to any top-level declaration in any file, which is a quick way of getting to the right file.
+
+![Open symbol by name](./open-symbol-by-name.gif)
 
 ---
