+++ src/extension.ts
@@ -8,6 +8,7 @@ import * as OfflinePackageDocs from "./features/offline-package-docs"
 import * as TypeDrivenAutocomplete from './features/type-driven-autocomplete'
 import * as HtmlToElm from './features/html-to-elm'
 import * as GoToSymbol from "./features/go-to-symbol"
+import * as OpenSymbolByName from "./features/open-symbol-by-name"
 
 export async function activate(context: vscode.ExtensionContext) {
   console.info("ACTIVATE")
@@ -34,6 +35,7 @@ export async function activate(context: vscode.ExtensionContext) {
   TypeDrivenAutocomplete.feature({ globalState, context })
   HtmlToElm.feature({ globalState, context })
   GoToSymbol.feature({ globalState, context })
+  OpenSymbolByName.feature({ globalState, context })
 }
 
 export function deactivate() {
